Renamed a file to financials.js, which includes a lot of different axios financial functions/calls for APIs. Continuing work on balance stuff.
