+++ client/src/actions/financials.js
@@ -1,0 +1,36 @@
+import axios from "axios";
+
+import { GET_ERRORS } from "./types";
+
+// Obtains transaction history from the user
+export const obtainTransHistory = () => dispatch =>
+{
+    axios
+        .post("/api/accounts/transactionhistory")
+        .then(res =>
+            {
+                
+            })
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
+};
+
+export const obtainBalance = (userID) => dispatch =>
+{
+    axios
+        .post("/api/accounts/findtippee", userID)
+        .then(res =>
+            {
+                console.log(res)
+            })
+        .catch(err =>
+            dispatch({
+                type: GET_ERRORS,
+                payload: err.response.data
+            })
+        );
+};
